Refactoring: moved some methods to separate classes in the Domain project
