+++ Amigula.Domain.Test/GamesServiceTest.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Amigula.Domain.Classes;
 using Amigula.Domain.DTO;
 using Amigula.Domain.Interfaces;
 using Amigula.Domain.Services;
@@ -84,102 +83,6 @@ namespace Amigula.Domain.Test
             Assert.IsNotNull(result);
             Assert.IsInstanceOfType(result, typeof (IEnumerable<GamesDto>));
             Assert.AreEqual(result.Count(), 3);
-        }
-
-        [TestMethod]
-        public void PrepareTitleScreenshot_GameTitle_ReturnsGameScreenshotsDto()
-        {
-            const string gameTitle = "Apidya v1.2 (1990) [Publisher name]";
-
-            var result = _gamesService.PrepareTitleScreenshot(gameTitle);
-
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, typeof (GameScreenshotsDto));
-            Assert.AreEqual(result.Title, "Apidya");
-            Assert.AreEqual(result.Screenshot1, "Apidya.png");
-            Assert.AreEqual(result.Screenshot2, "Apidya_1.png");
-            Assert.AreEqual(result.Screenshot3, "Apidya_2.png");
-            Assert.AreEqual(result.GameFolder, "A\\");
-        }
-
-        [TestMethod]
-        public void PrepareTitleScreenshot_GameTitleWithSpaces_ReturnsGameScreenshotsDto()
-        {
-            const string gameTitle = "International Karate Plus v1.3 (1988) [Publisher name]";
-
-            var result = _gamesService.PrepareTitleScreenshot(gameTitle);
-
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, typeof (GameScreenshotsDto));
-            Assert.AreEqual(result.Title, "International Karate Plus");
-            Assert.AreEqual(result.Screenshot1, "International_Karate_Plus.png");
-            Assert.AreEqual(result.Screenshot2, "International_Karate_Plus_1.png");
-            Assert.AreEqual(result.Screenshot3, "International_Karate_Plus_2.png");
-            Assert.AreEqual(result.GameFolder, "I\\");
-        }
-
-        [TestMethod]
-        public void PrepareTitleScreenshot_NumericGameTitle_ReturnsGameScreenshotsDto()
-        {
-            const string gameTitle = "1942 v1.3 (1988) [Publisher name]";
-
-            var result = _gamesService.PrepareTitleScreenshot(gameTitle);
-
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, typeof (GameScreenshotsDto));
-            Assert.AreEqual(result.Title, "1942");
-            Assert.AreEqual(result.Screenshot1, "1942.png");
-            Assert.AreEqual(result.Screenshot2, "1942_1.png");
-            Assert.AreEqual(result.Screenshot3, "1942_2.png");
-            Assert.AreEqual(result.GameFolder, "0\\");
-        }
-
-        [TestMethod]
-        public void PrepareTitleScreenshot_Null_ReturnsGameScreenshotsDto()
-        {
-            var result = _gamesService.PrepareTitleScreenshot(null);
-
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, typeof (GameScreenshotsDto));
-            Assert.IsNull(result.Title);
-            Assert.IsNull(result.Screenshot1);
-            Assert.IsNull(result.Screenshot2);
-            Assert.IsNull(result.Screenshot3);
-            Assert.IsNull(result.GameFolder);
-        }
-
-        [TestMethod]
-        public void PrepareTitleUrl_Title_ReturnsString()
-        {
-            const string gameTitle = "Apidya v1.2 (1990) [Publisher name]";
-
-            var result = _gamesService.PrepareTitleUrl(gameTitle);
-
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, typeof(string));
-            Assert.AreEqual(result, "Apidya");
-        }
-
-        [TestMethod]
-        public void PrepareTitleUrl_TitleWithSpaces_ReturnsStringWithoutSpaces()
-        {
-            const string gameTitle = "International Karate Plus v1.3 (1988) [Publisher name]";
-
-            var result = _gamesService.PrepareTitleUrl(gameTitle);
-
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, typeof(string));
-            Assert.AreEqual(result, "International%20Karate%20Plus");
-        }
-
-        [TestMethod]
-        public void PrepareTitleUrl_Null_ReturnsEmptyString()
-        {
-            var result = _gamesService.PrepareTitleUrl(null);
-
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, typeof(string));
-            Assert.AreEqual(result, "");
         }
 
         [TestMethod]
@@ -377,19 +280,6 @@ namespace Amigula.Domain.Test
             var enumerable = result as IList<string> ?? result.ToList();
             Assert.AreEqual(enumerable.Count(), 11);
             Assert.IsTrue(enumerable.Last().Equals(lastGameDisk));
-        }
-
-        [TestMethod]
-        public void AddGameScreenshot_GameFilename_ReturnsOperationResult()
-        {
-            const string screenshot = "Screenshot1.png";
-            const string gameTitle = "Apidya";
-
-            var result = _gamesService.AddGameScreenshot(gameTitle, screenshot);
-
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, typeof(OperationResult));
-            Assert.IsTrue(result.Success);
         }
     }
 }
