+++ Amigula.Domain.Test/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="AutoMapper" version="3.3.1" targetFramework="net452" userInstalled="true" />
-  <package id="FakeItEasy" version="1.25.2" targetFramework="net452" userInstalled="true" />
+  <package id="AutoMapper" version="3.3.1" targetFramework="net452" />
+  <package id="FakeItEasy" version="1.25.3" targetFramework="net452" />
 </packages>
