+++ Amigula.Domain/Amigula.Domain.csproj
@@ -11,6 +11,7 @@
     <AssemblyName>Amigula.Domain</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
+    <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -51,6 +52,8 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\EmulatorService.cs" />
     <Compile Include="Services\GamesService.cs" />
+    <Compile Include="Services\GameTitleService.cs" />
+    <Compile Include="Services\ScreenshotsService.cs" />
     <Compile Include="Services\YoutubeService.cs" />
   </ItemGroup>
   <ItemGroup />
